requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.0.3
idna==2.8
python-coveralls==2.9.1
PyYAML==5.1
requests==2.21.0
scriptinep3==0.3.1
six==1.12.0
urllib3==1.24.2
wrapt==1.11.1
